// src/core/http/generated/models/paymentGatewayStatus.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Petvex Admin API
 * OpenAPI spec version: 1.0.0
 */

export type PaymentGatewayStatus = typeof PaymentGatewayStatus[keyof typeof PaymentGatewayStatus];


export const PaymentGatewayStatus = {
  active: 'active',
  inactive: 'inactive',
  testing: 'testing',
  error: 'error',
} as const;